docfit-tool/requirements.txt
pymupdf>=1.24.0
pikepdf>=8.0.0
Pillow>=10.0.0
click>=8.1.0
numpy>=1.26.0
